finance_manager-original-main/finance_manager-original/fin_backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
SQLAlchemy==2.0.35
passlib==1.7.4
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
joblib==1.5.3
scikit-learn==1.6.1
